refactor: move createConfigurations to utils-os, eliminate manual prefix handling
All callers now go through createConfigurations().es() which guarantees
the prefix is always applied. No caller can forget it.
