+++ packages/api-headless-cms-pg-os/src/operations/entry/search/getUniqueFieldValues.ts
@@ -1,7 +1,7 @@
 import WebinyError from "@webiny/error";
 import type { OpenSearchSearchResponse } from "@webiny/api-opensearch";
 import { shouldIgnoreEsResponseError } from "@webiny/api-headless-cms-utils-os/operations/entry/elasticsearch/shouldIgnoreEsResponseError.js";
-import { getOpenSearchIndexPrefix } from "@webiny/api-opensearch";
+import { createConfigurations } from "@webiny/api-headless-cms-utils-os/configurations.js";
 import type { SearchOperationDeps } from "./types.js";
 import type { GetUniqueFieldValuesStorageOperation } from "@webiny/api-headless-cms/features/shared/storageOperations/entry/GetUniqueFieldValuesStorageOperation.js";
 
@@ -10,9 +10,8 @@ export const createGetUniqueFieldValuesOperation = (
 ): GetUniqueFieldValuesStorageOperation.Interface["execute"] => {
     return async (model, uniqueFieldValuesParams) => {
         const { where, fieldId } = uniqueFieldValuesParams;
-        const { index: rawIndex } = await deps.indexProvider.execute({ model });
-        const prefix = getOpenSearchIndexPrefix();
-        const index = prefix ? prefix + rawIndex : rawIndex;
+        const configurations = createConfigurations(deps.indexProvider);
+        const { index } = await configurations.es({ model });
 
         const field = model.fields.find(f => f.fieldId === fieldId);
         if (!field) {
